requirements.txt
pyserial==3.5
matplotlib~=3.8.4